### 2025-01-17 #7683 @editorialbot
[PRE REVIEW]: C4DYNAMICS: Python Framework for Dynamic Systems
**Submitting author:** @c4dynamics (<a href="http://orcid.org/0000-0002-7932-0562">Ziv Meri</a>)
**Repository:** https://github.com/C4dynamics/C4dynamics
**Branch with paper.md** (empty if default branch): main
**Version:** v2.0.0
