### 2025-01-21 #7683 @C4dynamics
Hi Afron, thanks.

I wanted to have my paper accepted toward a conference we have in May.

Is there anything I could do to prioritize my submission?

Thank you, Ziv (c4dynamics).
