### 2025-03-22 8bac300
user guide -> use cases. move the programs folder
