+++ programs/drifting_car.py
@@ -1,464 +1,0 @@
-# from scipy.integrate import odeint 
-# from scipy.linalg import solve_discrete_are
-import numpy as np 
-import cv2 
-from matplotlib import pyplot as plt 
-
-
-# not for demonstration: 
-import os, sys
-from enum import Enum  
-import zlib 
-
-
-
-
-plt.style.use('dark_background')  
-plt.switch_backend('TkAgg')
-
-
-def rootdir(dir):
-  print(dir)
-  if dir[-2:] == ':\\': return dir  
-  return rootdir(os.path.dirname(dir))
-
-def c4dir(dir, addpath = ''):
-  # dirname and basename are supplamentary:
-  # c:\dropbox\c4dynamics\text.txt
-  # dirname: c:\dropbox\c4dynamics
-  # basename: text.txt 
-  
-
-
-  inc4d = os.path.basename(dir) == 'c4dynamics'
-  hasc4d = any(f == 'c4dynamics' for f in os.listdir(dir) 
-                if os.path.isdir(os.path.join(dir, f)))
-
-
-  if inc4d and hasc4d: 
-    addpath += ''
-    return addpath
-  
-  addpath += '..\\'
-  return c4dir(os.path.dirname(dir), addpath)
-
-
-# rootdir(os.getcwd()) 
-
-print(os.getcwd())
-c4path = c4dir(os.getcwd())
-print(c4path)
-sys.path.append(c4path)
-
-
-import c4dynamics as c4d 
-from c4dynamics.utils.tictoc import tic, toc 
-
-savedir = os.path.join(c4path, 'docs', 'source', 'programs', 'figures', 'driftcar') 
-
-
-
-
-import pickle 
-
-yolo3 = c4d.detectors.yolov3()  
-
-
-# https://www.pexels.com/@abed-ismail/
-dname = 'drifting_car.mp4'
-vidpath = c4d.datasets.video('drifting_car')
-
-
-rtdetect = False # True # 
-
-if not rtdetect: 
-  
-  ptspath = os.path.join('examples', '_out')
-  ptsfile = os.path.join(ptspath, os.path.basename(vidpath)[:-4] + '.pkl')
-  
-  if os.path.exists(ptsfile):
-    with open(ptsfile, 'rb') as file:
-      detections = pickle.load(file)
-  else: 
-    detections = c4d.pixelpoint.video_detections(vidpath, storepath = ptspath)
-
-#
-video_cap = cv2.VideoCapture(vidpath)
-fps = video_cap.get(cv2.CAP_PROP_FPS)
-dt_frame = 1 / fps # 1 / frame per second = the length of a single frame
-N = int(video_cap.get(cv2.CAP_PROP_FRAME_COUNT)) # total frames count 
-tf = N * dt_frame 
-# 
-
-
-samplefactor = 2 
-
-
-
-A = np.zeros((6, 6))
-A[0, 4] = A[1, 5] = 1
-
-C = np.zeros((4, 6))
-C[0, 0] = C[1, 1] = C[2, 2] = C[3, 3] = 1
-
-
-# observability test 
-obsv = C
-n = len(A)
-for i in range(1, n):
-  obsv = np.vstack((obsv, C @ np.linalg.matrix_power(A, i)))
-rank = np.linalg.matrix_rank(obsv)
-c4d.cprint(f'The system is observable (rank = n = {n}).' if rank == n else f'The system is not observable (rank = {rank}, n = {n}).', 'y')
-
-dt = dt_frame / samplefactor # .0333 / 2 = .0166s 
-
-
-def steadystate():
-
-
-  c4d.cprint('steadystate', 'y')
-
[…]
