+++ README.md
@@ -34,6 +34,8 @@ Tsipor (bird) Dynamics (c4dynamics) is the open-source framework of algorithms d
 
 ✅ Built-in functions for Kalman filters
 
+✅ Integrated sensors and object detection models
+
 ✅ Out-of-the-box environments for reinforcement learning 
 
 ✅ Seamless integration with OpenCV & Open3D 
