### 2025-04-23 #7683 @kyleniemeyer
@C4dynamics per documentation, the proper way to specify affiliations is:
```
authors:
  - name: Adrian M. Price-Whelan
    orcid: 0000-0000-0000-0000
    equal-contrib: true
    affiliation: "1, 2" # (Multiple affiliations must be quoted)
  - name: Author Without ORCID
    equal-contrib: true # (This is how you can denote equal contributions between multiple authors)
    affiliation: 2
  - name: Author with no affiliation
    corresponding: true # (This is how to denote the corresponding author)
    affiliation: 3
  - given-names: Ludwig
    dropping-particle: van
    surname: Beethoven
    affiliation: 3
affiliations:
 - name: Lyman Spitzer, Jr. Fellow, Princeton University, United States
   index: 1
   ror: 00hx57361
 - name: Institution Name, Country
   index: 2
 - name: Independent Researcher, Country
   index: 3
```

so in your case, the author and affiliation block should change to something like:
```
authors:
  - name: Ziv Meri
    orcid: 0000-0002-7932-0562
    affiliation: 1
affiliations:
 - name: C4DYNAMICS
   index: 1
```

(However, normally the affiliation would provide an institution or company, and country/location.)
