### 2025-05-04 #7683 @C4dynamics
Hi @kyleniemeyer , just fixed it. PDF seems to build fine now.
