### 2025-06-29 #7683 @hweifluids
Dear Dr. Beck @phibeck and Dr Meri @C4dynamics ,

I have reviewed the code, the documentation, and the short paper submitted to JOSS. Below are my suggestions based on the JOSS checklist in the guideline:

√ **Repo and license.** The repository is hosted on GitHub with open access and includes an appropriate open-source license.
× **Authorship and affiliation.** The author, Dr Meri, is not correctly affiliated. I believe C4Dynamics is the name of the library, not the affiliation. Unless it represents a group of scholars, even a virtual one, C4Dynamics should not be listed as the affiliation. Please also verify whether the author list is complete or if this is a solo work.
√ **Installation.** Installation via pip is straightforward and completes in seconds.
√ **Functions and performance.** The library offers valuable tools for dynamic analysis, particularly in aerospace and robotics. Although it does not use MPI, its performance meets application needs. Actually speaking, I would really like to use it in my recent research on dynamic stability of HWB aircraft (before this, I planned to let students to code the case).
√ **Documentation.** The documentation is well organized, with several examples and detailed explanations. Btw, I really like the cute sketches in the tutorials.
√ **Automated testing.** Enabled.
× **Community guidelines.** Not found.
× **JOSS Manuscript.** The paper does not outline the background of dynamic modeling and simulation. In the Statement of Need, there is no clear research gap explaining why peers require this library. Related works are not properly cited or discussed, e.g., _SimuPy_ lib. Reference 01 is your own unpublished work and should not be cited. References 02 and 03 describe only your library’s requirements, not other works that could provide context. Please refer to a JOSS example paper to improve yours.

Therefore, I recommend that @phibeck rates this work (code and paper) as “minor revision”. It demonstrates strong academic value and potential, but requires improvements in readability and clarity, so that readers can benefit from the toolkit in a short time in their own research after reading the paper. After revision in a proper way, I believe it could meet the digital publication requirements of JOSS journal.

Best regards,
Huanxia
