### 2025-06-30 #7683 @phibeck
> Dear Dr. Beck,
>
> Thank you for the opportunity to review the code and manuscript submitted to JOSS. It would be my pleasure to undertake this review, and I will submit my report after reading the guideline and completing review asap.
>
> Best regards, Huanxia @ UoManchester

Thank you very much for agreeing to review @Huanxia! We are actually still in the [pre-review] issue and the official review hasn't yet started - I need to find one or two more reviewers until we're ready. Thank you though for starting your review so promptly! Once the official review starts (you'll get pinged), would you mind creating your checklist as instructed then (`@editorialbot generate my checklist`) and go through these points once more based on your assessment you already have? I hope to find another review asap. Thank you very much, and again thanks for taking care of your review so swiftly!
