### 2025-08-08 #7683 @borgesaugusto
Hi! Yes, I would be happy to review this submission.
