### 2025-08-12 C4dynamics/C4dynamics#65 @borgesaugusto
[JOSS] Documentation
Hi! Here is a list of Documentation issues I found, towards the JOSS review: https://github.com/openjournals/joss-reviews/issues/8776
- [x] The folder "examples" has a "run_examples" ipynb, which throws an error, as I believe the notebooks have been moved to doc/programs.
- [x] In some of the ipynb of docs/programs the introductory figures are missing; they are in the HTML, however.
- [x] There should be a mention of how to seek support.
- [x] The Python version required is missing, I believe >3.8 is a requirement, but is there an upper limit? (for example, one commit "limits tests to 3.11).
- [x] It would be ideal to have a line in the "contributing" file, or in the documentation specifying how to run the unit tests  (python tests/run_unittests.py).
- [x] When running the tests (using the command above), I ran into an error in test_rigidbody.py and in test_animate.py, as it requires open3d to work, which is not a dependency of requirements.txt. Is there a way to avoid this? For example, making the test optional or the requirement mandatory.
Thank you!
