### 2025-08-12 #8776 @borgesaugusto
Hi! I have finished my review. I have opened three issues in C4dynamic issue tracker. They correspond to the unchecked elements of the checklist.
This are:
Authorship: https://github.com/C4dynamics/C4dynamics/issues/64
Documentation: https://github.com/C4dynamics/C4dynamics/issues/65
Software paper: https://github.com/C4dynamics/C4dynamics/issues/66

In summary, I believe that the work is well written, and the package seems solid and useful. However, some improvements can be made to the documentation.
More importantly, concerning the software paper, I believe that there should be a better description of previous work, and how the C4dynamics package fits into the existing software ecosystem. This entails a more developed "State of the field" section. I also think that a better decription of the software can be included in the "Statement of need" section of the paper, since there is still space. All in all, (particulary due to the changes to the "State of the field" section)  I believe that this would be a Major revision.

One of the [issues I opened was regarding the authorship](https://github.com/C4dynamics/C4dynamics/issues/64). I do not know what is the policy that JOSS applies for institutions, but I would like to let you know about this @phibeck .
