### 2025-08-12 #8776 @hweifluids
@phibeck Sorry, I did not check the 5 boxes as I think they are False from my side. I will certainly open issues for them in a single day. Generally, as I said in the pre-review section, the package is great, but JOSS paper not, therefore I will suggest a major revision, especially on the statement of gap. Related packages with similar functions are not properly referred to.
