### 2025-08-13 #8776 @C4dynamics
Hi @phibeck, @borgesaugusto, and @hweifluids,

Thank you very much for the time, care, and detail you’ve put into reviewing C4DYNAMICS. I truly appreciate the thoughtful feedback and the speed at which you’ve carried out the review.

I’ll go through your comments and address the issues you’ve raised, following your instructions closely.

On the authorship point — C4DYNAMICS is neither an institution nor a company; it’s simply the name of the package.

Best regards,
Ziv
