+++ contributing.md
@@ -1,14 +1,15 @@
 
 # Quickstart for Contributors 🚀
 
-* Add a use case to the [Programs section](https://c4dynamics.github.io/C4dynamics/programs/index.html) of the documentation: 
-* Press the 'Fork' button (top-right corner of the c4dynamics page)
-* Choose a classic simulation or an existing project as your base.
-* Employ state objects from c4dynamics.
-* Incorporate modules from the scientific library (e.g. sensors, filters, detectors).  
-* Create a Jupyter notebook and write your program.  
-* Keep it simple yet functional.   
-* Add comments and documentation to explain your approach.  
-* Ensure it runs without errors and produces meaningful results.  
-* Use c4dynamics tools (e.g., plotting functions) to visualize outputs.  
-* Submit your example as a pull request.  
+Here's how you can add a use case to the [Programs section](https://c4dynamics.github.io/C4dynamics/programs/index.html) of the documentation: 
+* Fork the repository (top-right corner of the [GitHub page](https://github.com/C4dynamics/C4dynamics))
+* Pick a starting point: choose a classic simulation or an existing project as your base.
+* Build with [state objects](https://c4dynamics.github.io/C4dynamics/api/States.html).
+* Extend with modules from the scientific library (e.g. [sensors](https://c4dynamics.github.io/C4dynamics/api/Sensors.html), [filters](https://c4dynamics.github.io/C4dynamics/api/Filters.html), [detectors](https://c4dynamics.github.io/C4dynamics/api/Detectors.html)).  
+* Write your program in a Jupyter notebook:
+- Keep it simple yet functional.   
+- Add comments and documentation to explain your approach.  
+* Test thoroughly: make sure it runs without errors and produces meaningful results.  
+* Visualize outputs: use c4dynamics plotting tools.  
+* Submit a pull request with your notebook and a short description of what it demonstrates.
+✨ That’s it! Your example will help other engineers learn and build faster.
