### 2025-09-06 #8776 @C4dynamics
Hi @phibeck,
Thanks for clarifying!

To set expectations, would completing the documentation updates by the end of September be a reasonable timeframe, or would that be considered too long for the review process?
