### 2025-09-22 138900f
change position, velocity attributes to Position, Velocity; rearrange all doctests in one function under c4dynamics.__init__; support testing for no open3d installation; replace eval() with getattr(); initial support for RL environemnts; allow minutes in tictoc;
