+++ c4dynamics/states/state.py
@@ -80,6 +80,9 @@ class state:
     
   '''
 
+  # TODO: define state space, i.e. define the state properties such
+  #       as scope, limits, continuouty. etc. 
+  #     probably basicaly for rl algos. 
 
   # Α α # Β β # Γ γ # Δ δ # Ε ε # Ζ ζ # Η η # Θ θ # Ι ι 
   # Κ κ # Λ λ # Μ μ # Ν ν # Ξ ξ # Ο ο # Π π # Ρ ρ # Σ σ/ς
@@ -97,11 +100,12 @@ class state:
           ('Mu', '\u039C'), ('Nu', '\u039D'), ('Xi', '\u039E'), ('Omicron', '\u039F'),
             ('Pi', '\u03A0'), ('Rho', '\u03A1'), ('Sigma', '\u03A3'), ('Tau', '\u03A4'),
               ('Upsilon', '\u03A5'), ('Phi', '\u03A6'), ('Chi', '\u03A7'), ('Psi', '\u03A8'), ('Omega', '\u03A9'))
-    # 
-  _reserved_keys = ('X', 'X0', 'P', 'V')
+    #
+
+  _reserved_keys = ('X', 'X0', 'P', 'V', 'Position', 'Velocity', 'norm', 'normalize', 'data', '_data', '_prmdata', '_didx')
 
 
-  def __init__(self, **kwargs):    
+  def __init__(self, **kwargs):
     # TODO enable providing type for the setter.X output.      
 
     # the problem with this is it cannot be used with seeker and other c4d objects
@@ -118,6 +122,7 @@ class state:
     # 3. the user must provide his implementations for poisiton velcity etc.. like used in the P() function that there i 
     #   encountered the problem.
 
+    # self._dtype = np.float32
     self._data = []    # for permanent class variables (t, x, y .. )
     self._prmdata = {} # for user additional variables 
 
@@ -152,6 +157,7 @@ class state:
     # but then i need to iterate all the examples and remove the print from state presentations 
     param_names = ", ".join(self._prmdata.keys())
 
+    # FIXME Parameters is wrong. because currently parameters are determined only by those which stored at least once with the storeparams method. 
     return (f"<state object>\n"
             f"State Variables: {self.__str__()}\n"
             f"Initial Conditions (X0): {self.X0}\n"
@@ -163,6 +169,44 @@ class state:
   #     raise AttributeError(f"{param} is a reserved key. Keys {self._reserved_keys} cannot use as parameter names.")
   #   else:
   #     super().__setattr__(param, value)
+
+  # @property 
+  # def dtype(self) -> np.dtype:
+  #   '''
+  #   Gets and sets the data type of the state variables. 
+
+  #   The data type is used to determine the type of the state variables. 
+  #   The default data type is :code:`np.float64`. 
+
+  #   Parameters
+  #   ----------
+  #   dtype : numpy.dtype 
+  #       Data type to set the state variables. 
+
+  #   Returns
+  #   -------
+  #   out : numpy.dtype 
+  #       Data type of the state variables. 
+
+  #   Examples
+  #   --------
+
+  #   .. code:: 
+
+  #     >>> s = c4d.state(x = 1, y = 0)
+  #     >>> s.dtype
+  #     dtype('float64')
+  #     >>> s.dtype = np.float32
+  #     >>> s.dtype
+  #     dtype('float32')
+
+  #   '''
+  #   return self._dtype
+
+  # @dtype.setter
+  # def dtype(self, dtype):
+  #   self._dtype = np.dtype(dtype)
+
 
 
 
@@ -233,15 +277,16 @@ class state:
       if k == 't': continue
       # the alteast_1d() + the flatten() is necessary to 
       # cope with non-homogenuous array 
[…]
