### 2025-09-22 C4dynamics/C4dynamics#65 @borgesaugusto
Thank you for the changes!
In point 2, I meant that if you access the ipynbs directly, you don't see the images (i.e https://github.com/C4dynamics/C4dynamics/blob/main/docs/programs/ballistic_coefficient.ipynb just below Cell 7)
Maybe this is "on purpose", as the docs folder is only used for compiling the github.io documentation?
