### 2025-09-22 C4dynamics/C4dynamics#65 @C4dynamics
@borgesaugusto
It is strange.
However, these notebooks are compiled with nbsphinx and designed to view under the use cases section of the documentation:
https://c4dynamics.github.io/C4dynamics/programs/ballistic_coefficient.html
Are you able to see the images here?
