### 2025-09-22 C4dynamics/C4dynamics#65 @borgesaugusto
Yes! now they are visible
