### 2025-09-25 #8776 @C4dynamics
By the way, who’s responsible for closing the issues once they’re fulfilled?
@phibeck
