### 2025-09-25 #8776 @phibeck
Thank you for the update. In practice it can be either author or reviewer to close the issue if they feel that the issue has been resolved - if that turns out not to be the case it can always be reopened.
