### 2025-09-26 69dd736
1. simplify the contributing.md. 2. move the pdf paper to the paper folder by github actions.
