+++ .github/workflows/paper.yml
@@ -21,7 +21,7 @@ jobs:
           # This is the output path where Pandoc will write the compiled
           # PDF. Note, this should be the same directory as the input
           # paper.md
-          path: paper/paper.pdf
+          path: paper/*.pdf
 
 
 
