### 2025-09-27 1d53f80
update paper and supporting files
