+++ paper/paper.bib
@@ -1,0 +1,57 @@
+% KALMAN  
+@ARTICLE{1104873,
+  author  = {Kalman, R.},
+  journal = {IRE Transactions on Automatic Control}, 
+  title   = {On the general theory of control systems}, 
+  year    = {1959},
+  volume  = {4},
+  number  = {3},
+  pages   = {110-110},
+  keywords= {Control systems;Noise measurement;Pollution measurement;Kalman filters;Optimal control;Wiener filter;Controllability;Observability},
+  doi     = {10.1109/TAC.1959.1104873}
+}
+
+% SIMUPY 
+@Article{Margolis2017simupy,
+  author  = {Margolis, Benjamin WL},
+  journal = {J. Open Source Software},
+  title   = {{SimuPy}: A {Python} framework for modeling and simulating dynamical systems},
+  year    = {2017},
+  number  = {17},
+  pages   = {396},
+  volume  = {2},
+  doi     = {10.21105/joss.00396},
+}
+
+% BDSIM
+@article{NEVAY2020107200,
+  title = {BDSIM: An accelerator tracking code with particle–matter interactions},
+  journal = {Computer Physics Communications},
+  volume = {252},
+  pages = {107200},
+  year = {2020},
+  issn = {0010-4655},
+  doi = {https://doi.org/10.1016/j.cpc.2020.107200},
+  url = {https://www.sciencedirect.com/science/article/pii/S0010465520300400},
+  author = {L.J. Nevay and S.T. Boogert and J. Snuverink and A. Abramov and L.C. Deacon and H. Garcia-Morales and H. Lefebvre and S.M. Gibson and R. Kwee-Hinzmann and W. Shields and S.D. Walker},
+  keywords = {Monte Carlo simulation, Particle accelerator, Geant4, Particle physics, Particle tracking},
+  abstract = {Beam Delivery Simulation (BDSIM) is a program that simulates the passage of particles in a particle accelerator. It uses a suite of standard high energy physics codes (Geant4, ROOT and CLHEP) to create a computational model of a particle accelerator that combines accurate accelerator tracking routines with all of the physics processes of particles in Geant4. This unique combination permits radiation and detector background simulations in accelerators where both accurate tracking of all particles is required over long range or over many revolutions of a circular machine, as well as interaction with the material of the accelerator.
+  Program summary
+  Program Title: BDSIM Program Files doi: http://dx.doi.org/10.17632/bzg5hc65h6.1 Licensing provisions: GNU General Public License 3 Programming language: C++, flex, bison External routines/libraries: Geant4, CLHEP, ROOT, gzstream, CMake Nature of problem: Simulate energy deposition and charged particle detector background in a particle accelerator originating from beam loss where particles may pass both through the vacuum pipe with magnetic and electromagnetic fields, as well as through the material of the magnets and accelerator itself. Simulate the passage of particles both through an accelerator and the surrounding material such as air. Do so in a sufficiently flexible way that a variety of accelerator configurations can be easily simulated. Solution method: Automatic creation of a 3D Geant4 model from an optical description of an accelerator using a library of generic 3D models that are user extendable. Accelerator tracking routines, the associated fields and coordinates transforms are provided for accurate magnetic field tracking.}
+}
+
+% ROBODART 
+@article{Chatzilygeroudis2024, 
+  doi = {10.21105/joss.06771}, 
+  url = {https://doi.org/10.21105/joss.06771}, 
+  year = {2024}, 
+  publisher = {The Open Journal}, 
+  volume = {9}, 
+  number = {102}, 
+  pages = {6771}, 
[…]
