### 2025-09-27 5a538f8
add nocite to paper.md
