### 2025-09-27 14fcbe4
build paper draft
