+++ paper/paper.md
@@ -70,7 +70,8 @@ plt.show()
 
 # References
 \nocite{1104873}
-\nocite{[@Margolis2017simupy]}
+<!-- \nocite{[@Margolis2017simupy]} -->
+\nocite{Margolis2017simupy}
 
 ---
 bibliography: paper.bib
