### 2025-09-27 9063a9d
paper draft
