### 2025-09-27 f831211
paper draft
