### 2025-09-27 58763ee
paper draft
