### 2025-09-27 225b4bd
paper draft + luenberger ref
