+++ paper/paper.bib
@@ -11,6 +11,20 @@
   doi     = {10.1109/TAC.1959.1104873}
 }
 
+
+% LUENBERGER
+@ARTICLE{3830260412,
+  author  = {David G. Luenberger},
+  publisher = {John Wiley & Sons}, 
+  title   = {Introduction to Dynamic Systems, Theory, Models, and Applications}, 
+  year    = {1979},
+  pages   = {446-446},
+  city    = {New York},
+  doi     = {10.1109/TAC.1959.1104873}
+}
+
+
+
 % SIMUPY 
 @Article{Margolis2017simupy,
   author  = {Margolis, Benjamin WL},
