### 2025-09-30 38ea510
revisit doc as issued by crhea93 reviewer from pyopensci
