+++ c4dynamics/states/lib/__init__.py
@@ -1,282 +1,20 @@
 '''
-This page is an `introduction` to the states library. 
-For the different pre-defined states themselves, go to :ref:`states-header`.     
 
-.. currentmodule:: c4dynamics.states.lib
+Pre-defined state objects
 
-Each one of the states in the library is inherited from the 
-:class:`state <c4dynamics.states.state.state>` 
-class and has the benefit of its attributes, like 
-:meth:`store() <c4dynamics.states.state.state.store>` 
-:meth:`data() <c4dynamics.states.state.state.data>` 
-etc. 
 
+.. list-table:: 
+  :header-rows: 0
 
-1. Data Point 
--------------
-
-C4dynamics provides built-in entities for developing 
-and analyzing algorithms of objects in space and time:
-
-:class:`datapoint <datapoint.datapoint>`: 
-a class defining a point in space: position, velocity, and mass.
-
-:class:`rigidbody <rigidbody.rigidbody>`: 
-a class rigidbody a class defining a rigid body in space, i.e. 
-an object with length and angular position.
-
-
-.. figure:: /_architecture/body_states.svg
-  :width: 482px
-  :height: 534px   
-
-  **Figure:** 
-  Conceptual diagram showing the relationship between the two 
-  fundamental objects used to describe bodies in space: 1) the
-  datapoint, 2) the rigidbody. A rigidbody object extends the 
-  datapoint by adding on it body rotational motion. 
-
-
-The :class:`datapoint <datapoint.datapoint>` 
-is the most basic element in translational dynamics; it's a point in space. 
-
-A `datapoint` serves as the building block for modeling and simulating 
-the motion of objects in a three-dimensional space. 
-In the context of translational dynamics, a datapoint represents 
-a point mass in space with defined Cartesian coordinates :math:`(x, y, z)` 
-and associated velocities :math:`(v_x, v_y, v_z)`. 
-
-
-Data Attributes
-^^^^^^^^^^^^^^^ 
-
-State variables: 
-
-.. math:: 
-
-  X = [x, y, z, v_x, v_y, v_z]^T 
-
-- Position coordinates, velocity coordinates. 
-
-Parameters: 
-
-- `mass`: point mass.
-
-  
-Construction
-^^^^^^^^^^^^
-
-A `datapoint` instance is created by making a direct call to the datapoint constructor:
-
-.. code::
-
-  >>> from c4dynamics import datapoint 
-  >>> dp = datapoint()
-
-.. code::
-
-  >>> print(dp)
-  [ x  y  z  vx  vy  vz ]
-
-
-
-
-
-Initialization of an instance does not require any mandatory parameters. 
-However, 
-setting values to any of the state variables uses as initial conditions: 
-
-.. code::
-
-  >>> dp = datapoint(x = 1000, vx = -100)
-
-  
-Functionality 
-^^^^^^^^^^^^^  
-
-The :meth:`inteqm() <datapoint.datapoint.inteqm>` method uses 
-the Runge-Kutta integration technique 
-to evolve the state in response to external forces. 
-The mechanics underlying the equations of motion can be found 
-:mod:`here <c4dynamics.eqm>`. 
-
-The method :meth:`plot() <datapoint.datapoint.plot>` adds on 
-the standard :meth:`state.plot() <c4dynamics.states.state.state.plot>` 
-the option to draw trajectories from side view and from top view.  
-  
-
-
-
-See Also
-^^^^^^^^
-.datapoint
-.eqm
-
- 
[…]
