+++ c4dynamics/filters/__init__.py
@@ -1,1250 +1,17 @@
-"""
+'''
 
-This page is an `introduction` to the filters module. 
-For the different filter objects themselves, go to :ref:`filters-header`.     
+.. list-table:: 
+  :header-rows: 0
 
+  * - :class:`kalman <c4dynamics.filters.kalman.kalman>`
+    - Kalman filter
+  * - :class:`ekf <c4dynamics.filters.ekf.ekf>`
+    - Extended Kalman filter
+  * - :class:`lpf <c4dynamics.filters.lowpass.lowpass>`
+    - Lowpass filter
 
 
-The :mod:`filters <c4dynamics.filters>` module in `c4dynamics` provides a collection of 
-classes and functions for implementing various types of filters commonly used in control 
-systems and state estimation. 
-
-
-
-*******************
-Background Material
-*******************
-
-The background material and the sections concerning the particular filters 
-are based on sources in references [AG]_ [SD]_ [ZP]_.  
-
-System Model
-============
-
-State-space representation is a mathematical model of a physical system represented 
-as a set of input, output, and state variables related by first-order differential 
-(or difference) equations. 
-
-The state vector :math:`X` of a state-space model provides a snapshot of the system's current condition, 
-encapsulating all the variables necessary to describe its future behavior given the input.
-(In `c4dynamics` the state vector is a fundamental data structure, 
-represented by the class :class:`state <c4dynamics.states.state.state>`) 
-and a snapshot of its values is provided by 
-the property :attr:`state.X <c4dynamics.states.state.state.X>`). 
-
-When the coefficients of the state variables in the equations are constant, the 
-state model represents a linear system (LTI, linear time invariant). 
-If the coefficients are 
-linear functions of time, the system is considered linear time varying.  
-Otherwise, the system is nonlinear. 
-
-
-Nonlinear Systems
-=================
-
-All systems are naturally nonlinear. When an equilibrium point 
-representing the major operation part of the system can be found, then a 
-linearization is performed about this point, and the system is regarded 
-linear. 
-When such a point cannot be easily found, more advanced approaches 
-have to be considered to analyze and manipulate the system. Such 
-an approach is the extended Kalman filter. 
-
-A nonlinear system is described by:
-
-.. math::
-  :label: nonlinearmodel
-
-  \\dot{x}(t) = f(x, u) + \\omega 
-
-  y(t) = h(x) + \\nu 
-
-  x(0) = x_0 
-
-
-
-Where: 
-
-- :math:`f(\\cdot)` is an arbitrary vector-valued function representing the system process
-- :math:`x` is the system state vector 
-- :math:`t` is a time variable 
-- :math:`u` is the system input signal
-- :math:`\\omega` is the process uncertainty with covariance matrix :math:`Q_c`
-- :math:`y` is the system output vector (the measure)
-- :math:`h(\\cdot)` is an arbitrary vector-valued function representing the measurement equations 
-- :math:`\\nu` is the measure noise with covariance matrix :math:`R_c`
-- :math:`x_0` is a vector of initial conditions  
-
-The noise processes :math:`\\omega(t)` and :math:`\\nu(t)` are white, zero-mean, uncorrelated, 
-and have known covariance matrices :math:`Q_c` and :math:`R_c`, respectively:
-
-.. math::
-
-  \\omega(t) \\sim (0, Q_c) 
-
-  \\nu(t) \\sim (0, R_c) 
-
-  E[\\omega(t) \\cdot \\omega^T(t)] = Q_c \\cdot \\delta(t) 
-
[…]
