### 2025-10-01 3fa1660
paper: python lacks dedicated fw tailored ssmodeling
