### 2025-10-01 8748645
Update paper.md
