### 2025-10-01 95de6de
Update paper.md
