+++ paper/paper.md
@@ -10,6 +10,10 @@ tags:
 authors:
   - name: Ziv Meri
     orcid: 0000-0002-7932-0562
+    affiliation: 1
+affiliations:
+  - name: Independent 
+    index: 1    
 date: "2024-12-17"
 bibliography: paper.bib
 ---
