+++ c4dynamics/utils/plottracks.py
@@ -1,3 +1,6 @@
+# -*- coding: utf-8 -*-
+# pragma: no cover
+
 # import time 
 import os, sys 
 import argparse
