+++ c4dynamics/utils/cprint.py
@@ -56,27 +56,6 @@ def cprint(txt = '', color = 'white', end = '\n'):
 
 if __name__ == "__main__":
 
-  # import os 
-  # import doctest, contextlib
-  # from c4dynamics import IgnoreOutputChecker, cprint
-  
-  # # Register the custom OutputChecker
-  # doctest.OutputChecker = IgnoreOutputChecker
-
-  # tofile = False 
-  # optionflags = doctest.FAIL_FAST
-
-  # if tofile: 
-  #   with open(os.path.join('tests', '_out', 'output.txt'), 'w') as f:
-  #     with contextlib.redirect_stdout(f), contextlib.redirect_stderr(f):
-  #       result = doctest.testmod(optionflags = optionflags) 
-  # else: 
-  #   result = doctest.testmod(optionflags = optionflags)
-
-  # if result.failed == 0:
-  #   cprint("All tests passed!", 'g')
-  # else:
-  #   print(f"{result.failed}")
   from c4dynamics import rundoctests
   rundoctests(sys.modules[__name__])
 
