### 2025-10-02 C4dynamics/C4dynamics#65 @C4dynamics
Hi @borgesaugusto,

Thanks a lot for rerunning the tests and for your detailed feedback! I really appreciate it.

Regarding the points you mentioned:

1. YOLOv3 cfg file: I replaced the relative-path loading with importlib.resources, so the model configuration is now loaded in a way that works reliably regardless of the current working directory or installation method. This should prevent the Failed to open NetParameter file error in doctests.

2. NumPy C-extensions error: This type of error usually happens due to environment inconsistencies (e.g., mixing pip-installed packages inside a conda environment). In my environment, import numpy as np works fine. Using a clean Python 3.11 environment and installing the dependencies via the new pyproject.toml should help avoid this issue.

3. Installation from source: I added a pyproject.toml file so that installing the package is now simpler and only requires:

```pip install . ```

If dev dependencies are needed (for docs or tests), they can be installed with:

```pip install .[dev]```

This should eliminate the need for separate ```pip install -r requirements.txt ```steps.

It should now work out-of-the-box. Please dont hesitate to reach out for any issue.

Thanks again for your review and patience!

Best,
Ziv
