### 2025-10-02 a637eb6
exclude tex files from github counter, show the reference manus in the nav bar
