### 2025-10-13 f77d8f0
update paper affiliation, add autopilot example
