+++ paper/paper.md
@@ -11,6 +11,9 @@ authors:
   - name: Ziv Meri
     orcid: 0000-0002-7932-0562
     affiliation: 1
+affiliations:
+  - name: Independent Researcher, Israel
+    index: 1
 date: "2024-12-17"
 bibliography: paper.bib
 ---
