### 2025-10-22 9eee846
new installation guide
