### 2025-10-26 3f3bca6
fix references in filter concept page
