+++ paper/paper.bib
@@ -1,5 +1,5 @@
 % KALMAN  
-@ARTICLE{1104873,
+@ARTICLE{Kalman1104873,
   author  = {Kalman, R.},
   journal = {IRE Transactions on Automatic Control}, 
   title   = {On the general theory of control systems}, 
@@ -13,16 +13,67 @@
 
 
 % LUENBERGER
-@ARTICLE{3830260412,
+@ARTICLE{Luenberger3830260412,
   author  = {David G. Luenberger},
   publisher = {John Wiley & Sons}, 
   title   = {Introduction to Dynamic Systems, Theory, Models, and Applications}, 
   year    = {1979},
   pages   = {446-446},
   city    = {New York},
-  doi     = {10.1109/TAC.1959.1104873}
+  doi     = {https://doi.org/10.1002/bs.3830260412}
 }
 
+
+% NUMPY
+@ARTICLE{numpy2020,
+  author  = {Harris, C.R., Millman, K.J., van der Walt, S.J. et al},
+  journal = {Nature},
+  title   = {Array programming with NumPy},
+  year    = {2020},
+  volume  = {585},
+  number  = {7825},
+  pages   = {357-362},
+  doi     = {10.1038/s41586-020-2649-2},
+  url     = {https://doi.org/10.1038/s41586-020-2649-2}
+}
+
+
+% SCIPY
+@ARTICLE{SciPy2020,
+  author  = {Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E. and
+            Haberland, Matt and Reddy, Tyler and Cournapeau, David and
+            Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and
+            Bright, Jonathan and {van der Walt}, St{\'e}fan J. and
+            Brett, Matthew and Wilson, Joshua and Millman, K. Jarrod and
+            Mayorov, Nikolay and Nelson, Andrew R. J. and Jones, Eric and
+            Kern, Robert and Larson, Eric and Carey, C J and
+            Polat, {\.I}lhan and Feng, Yu and Moore, Eric W. and
+            {VanderPlas}, Jake and Laxalde, Denis and Perktold, Josef and
+            Cimrman, Robert and Henriksen, Ian and Quintero, E. A. and
+            Harris, Charles R. and Archibald, Anne M. and
+            Ribeiro, Ant{\^o}nio H. and Pedregosa, Fabian and
+            {van Mulbregt}, Paul and {SciPy 1.0 Contributors}},
+  title   = {{{SciPy} 1.0: Fundamental Algorithms for Scientific
+            Computing in Python}},
+  journal = {Nature Methods},
+  year    = {2020},
+  volume  = {17},
+  pages   = {261--272},
+  adsurl  = {https://rdcu.be/b08Wh},
+  doi     = {10.1038/s41592-019-0686-2},
+}
+
+
+% ROBOTICS SOFTWARE: PAST, PRESENT, AND FUTURE
+@ARTICLE{Gerkey2003robotics,
+  author  = {Gerkey, B.P. and Vaughan, R.T. and Howard, A.},
+  journal = {Proceedings of the IEEE International Conference on Robotics and Automation (ICRA)},
+  title   = {Robotics Software: Past, Present, and Future},
+  year    = {2003},
+  pages   = {2145-2150},
+  doi     = {10.1109/ROBOT.2003.1242018},
+  url     = {https://doi.org/10.1109/ROBOT.2003.1242018}
+}
 
 
 % SIMUPY 
@@ -36,6 +87,7 @@
   volume  = {2},
   doi     = {10.21105/joss.00396},
 }
+
 
 % BDSIM
 @article{NEVAY2020107200,
@@ -54,6 +106,7 @@
[…]
