### 2025-11-08 8d9a35e
more paper edits
