### 2025-11-12 #8776 @borgesaugusto
Hi! I don't have any further comments on this submission. It is a really nice package, and the author has made a great effort to improve both the code and its presentation.

I believe this completes my review @phibeck
