### 2025-11-19 8e16bb2
fixes a critical bug in the core `State` object assignment mechanism
