+++ c4dynamics/states/state.py
@@ -133,13 +133,25 @@ class state:
     self._prmdata = {} # for user additional variables 
 
     self._didx = {'t': 0}
+    
+    # if any of the input vars is floating point, then it casts the entire state type (no partial types array)
+    # if all of them are integer, it casts the staet type. 
+    # i think maybe the best way to achieve it is by generating a numpy array and check its dtype: 
+    self._X = np.array(list(kwargs.values()))
+    # self._X = np.zeros(len(kwargs)) #, dtype = np.float64)
+    # TODO: now the type is determined by the initial values. 
+    #       if any is floating point then the entire state is float.
+    #       otherwise, integer.
+    #       however, there cannot be partial types such as some variables. 
+    #       this must be covered in the documentation.
 
     for i, (k, v) in enumerate(kwargs.items()):
       if k in self._reserved_keys:
         raise ValueError(f"{k} is a reserved key. Keys {self._reserved_keys} cannot use as variable names.")
-      setattr(self, k, v)
+      # setattr(self, k, v)
       setattr(self, k + '0', v)
       self._didx[k] = 1 + i
+      self._X[i] = v
 
 
   def __str__(self):
@@ -170,48 +182,7 @@ class state:
             f"Current State Vector (X): {self.X}\n"
             f"Parameters: {param_names if param_names else 'None'}")
 
-  # def __setattr__(self, param, value):
-  #   if param in self._reserved_keys:
-  #     raise AttributeError(f"{param} is a reserved key. Keys {self._reserved_keys} cannot use as parameter names.")
-  #   else:
-  #     super().__setattr__(param, value)
-
-  # @property 
-  # def dtype(self) -> np.dtype:
-  #   '''
-  #   Gets and sets the data type of the state variables. 
-
-  #   The data type is used to determine the type of the state variables. 
-  #   The default data type is :code:`np.float64`. 
-
-  #   Parameters
-  #   ----------
-  #   dtype : numpy.dtype 
-  #       Data type to set the state variables. 
-
-  #   Returns
-  #   -------
-  #   out : numpy.dtype 
-  #       Data type of the state variables. 
-
-  #   Examples
-  #   --------
-
-  #   .. code:: 
-
-  #     >>> s = c4d.state(x = 1, y = 0)
-  #     >>> s.dtype
-  #     dtype('float64')
-  #     >>> s.dtype = np.float32
-  #     >>> s.dtype
-  #     dtype('float32')
-
-  #   '''
-  #   return self._dtype
-
-  # @dtype.setter
-  # def dtype(self, dtype):
-  #   self._dtype = np.dtype(dtype)
+  
 
 
 
@@ -219,6 +190,32 @@ class state:
   #
   # state operations
   ## 
+
+
+  def __getattr__(self, var):
+    if var in self._didx:
+      return self._X[self._didx[var] - 1]
+    else:
+      # default Python behavior
+      raise AttributeError(f"{var} not found")
+
+
+
+  def __setattr__(self, var, val):
+    # intercept known state variables
+    if "_didx" in self.__dict__ and var in self._didx:
+
+      if np.array(val).dtype != self._X.dtype:
+        warnings.warn(
+              f"Type mismatch! Casting X to {np.array(val).dtype}. (currently: {self._X.dtype})."
+          , c4d.c4warn)
[…]

[…]
