### 2025-12-01 f1c3a7d
his update replaces the behavior from the previous release, where a dtype mismatch triggered a warning.
