+++ examples/filters/kf/intro.py
@@ -130,15 +130,16 @@ def steadystate(cont = False):
   # tgt.v = 1
   v = 150
   sensor_noise = 200 
-  if cont: 
-    kf = kalman({'x': 0.}, P0 = 0.5**2, A = 0, B = v, C = 1, Q = 0.05, R = sensor_noise**2, steadystate = True, dt = 1)
-  else: 
-    kf = kalman({'x': 0.}, P0 = 0.5**2, F = 1, G = v, H = 1, Q = 0.05, R = sensor_noise**2, steadystate = True)
+  # if cont: 
+  #   kf = kalman({'x': 0.}, P0 = 0.5**2, A = 0, B = v, C = 1, Q = 0.05, R = sensor_noise**2, steadystate = True, dt = 1)
+  # else: 
+  kf = kalman({'x': 0.}, P0 = 0.5**2, F = 1, G = v, H = 1, Q = 0.05, R = sensor_noise**2, steadystate = True)
     
 
   
 
   train = c4d.state(x = 0.)
+  train.x = 0
 
   
   for t in range(1, 25 + 1): #  seconds. 
