### 2025-12-02 f97ceb0
update version to 2.3.1 and adjust Python compatibility in setup files
