+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "c4dynamics"
-version = "2.3.0"
+version = "2.3.1"
 description = "Python framework for state-space modeling and algorithm development"
 authors = [
     {name = "Ziv Meri", email="zivmeri@gmail.com"}
