### 2025-12-12 #8776 @C4dynamics
Hi @phibeck, thank you for your detailed comments!
I wanted to let you know that I’ll be unavailable for the next ~24-48 hours.
Right after that, I will carefully review all your notes (including the Zenodo version tag, DOI fixes, archive title, and author list differences) and respond accordingly.

Thank you for your patience!
